1. 数据库结构改了：task_record.task_name 外键到 task.task_name， 2. 增加 /abnormal/record/selectAbnormalRecordDetail 根据 task_record.id 返回abnormalRecordDetail[包括了abnormal,abnormal_image,abnormal_recorde]。3. abnormal,abnormal_image，task_record,user 等支持add/update.

+++ src/main/java/com/eservice/api/model/abnormal/Abnormal.java
@@ -5,7 +5,7 @@ import javax.persistence.*;
 public class Abnormal {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Byte id;
+    private Integer id;
 
     /**
      * 异常名称
@@ -16,14 +16,14 @@ public class Abnormal {
     /**
      * @return id
      */
-    public Byte getId() {
+    public Integer getId() {
         return id;
     }
 
     /**
      * @param id
      */
-    public void setId(Byte id) {
+    public void setId(Integer id) {
         this.id = id;
     }
